Add cache disabling mechanism and refactor UpdateEntityAsync
Introduced a `DisableCache` method in the `IRepositoryDecoratorBase<TEntity, TDto>` interface and implemented it in the `RepositoryDecoratorBase<TEntity, TDto>` class. This method sets a private `_disabledCache` field to `true`. The `IsCacheEnabled` property now checks `_disabledCache` in addition to `CacheDurationMinutes > 0`. Also, moved `repoObj` retrieval and null check inside the `try` block in `UpdateEntityAsync` for better transaction handling.